feat: extrapolated the logic for fetching the stored SiteSettings in another class
